feat: ethereum fee context for ckEth
