+++ src/frontend/src/icp/components/send/IcFeeDisplay.svelte
@@ -12,4 +12,4 @@
 
 <BitcoinEstimatedFee />
 <BitcoinKYTFee {networkId} />
-<EthereumEstimatedFee {networkId} />
+<EthereumEstimatedFee />
